#60 optimization front

+++ passenger-transportation-service/passenger-transportation-jpa-graph-adapter/src/main/java/org/service/output_port/util/EdgeUtils.java
@@ -26,7 +26,7 @@ public class EdgeUtils {
         Map<Integer, Edge> longEdgeMap = new HashMap<>();
 
         for (var j : a) {
-            if (!longEdgeMap.containsKey(j.getId())) {
+            if (!longEdgeMap.containsKey(j.getId().intValue())) {
                 longEdgeMap.put(Math.toIntExact(j.getId()), j);
             }
         }
